+++ VRDR.Tests/Messaging_Should.cs
@@ -32,8 +32,8 @@ namespace VRDR.Tests
             Assert.NotNull(submission.MessageTimestamp);
             Assert.Null(submission.MessageSource);
             Assert.NotNull(submission.MessageId);
-            Assert.Null(submission.CertificateNumber);
-            Assert.Null(submission.StateAuxiliaryIdentifier);
+            Assert.Null(submission.CertNo);
+            Assert.Null(submission.StateAuxiliaryId);
             Assert.Null(submission.NCHSIdentifier);
         }
 
@@ -46,9 +46,9 @@ namespace VRDR.Tests
             Assert.Equal("2019-02-20T16:48:06-05:00", submission.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", submission.MessageType);
             Assert.Equal(10,10);
-            Assert.Equal((uint)182, submission.CertificateNumber);
+            Assert.Equal((uint)182, submission.CertNo);
             Assert.Equal((uint)2019, submission.DeathYear);
-            Assert.Equal("000000000042", submission.StateAuxiliaryIdentifier);
+            Assert.Equal("000000000042", submission.StateAuxiliaryId);
             Assert.Equal("2019YC000182", submission.NCHSIdentifier);
 
             record = (DeathRecord)JSONRecords[0];
@@ -56,17 +56,17 @@ namespace VRDR.Tests
             Assert.NotNull(submission.DeathRecord);
             Assert.Equal("2018-02-20T16:48:06-05:00", submission.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", submission.MessageType);
-            Assert.Equal((uint)182, submission.CertificateNumber);
+            Assert.Equal((uint)182, submission.CertNo);
             Assert.Equal((uint)2019, submission.DeathYear);
-            Assert.Equal("000000000042", submission.StateAuxiliaryIdentifier);
+            Assert.Equal("000000000042", submission.StateAuxiliaryId);
             Assert.Equal("2019YC000182", submission.NCHSIdentifier);
 
             record = null;
             submission = new DeathRecordSubmissionMessage(record);
             Assert.Null(submission.DeathRecord);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", submission.MessageType);
-            Assert.Null(submission.CertificateNumber);
-            Assert.Null(submission.StateAuxiliaryIdentifier);
+            Assert.Null(submission.CertNo);
+            Assert.Null(submission.StateAuxiliaryId);
             Assert.Null(submission.NCHSIdentifier);
 
             record = (DeathRecord)JSONRecords[1];
@@ -74,8 +74,8 @@ namespace VRDR.Tests
             Assert.NotNull(submission.DeathRecord);
             Assert.Equal("2019-02-20T16:48:06-05:00", submission.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", submission.MessageType);
-            Assert.Null(submission.CertificateNumber);
-            Assert.Null(submission.StateAuxiliaryIdentifier);
+            Assert.Null(submission.CertNo);
+            Assert.Null(submission.StateAuxiliaryId);
             Assert.Null(submission.NCHSIdentifier);
         }
 
@@ -86,16 +86,16 @@ namespace VRDR.Tests
             Assert.Equal("2018-02-20T16:48:06-05:00", submission.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", submission.MessageType);
             Assert.Equal("2018MA000001", submission.NCHSIdentifier);
-            Assert.Equal((uint)1, submission.CertificateNumber);
+            Assert.Equal((uint)1, submission.CertNo);
             Assert.Equal((uint)2018, submission.DeathYear);
-            Assert.Equal("42", submission.StateAuxiliaryIdentifier);
-            Assert.Equal(submission.DeathJurisdictionID, submission.DeathRecord.DeathLocationJurisdiction);
+            Assert.Equal("42", submission.StateAuxiliaryId);
+            Assert.Equal(submission.JurisdictionId, submission.DeathRecord.DeathLocationJurisdiction);
 
             submission = BaseMessage.Parse<DeathRecordSubmissionMessage>(FixtureStream("fixtures/json/DeathRecordSubmissionNoIdentifiers.json"));
             Assert.Equal("2018-02-20T16:48:06-05:00", submission.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", submission.MessageType);
-            Assert.Null(submission.CertificateNumber);
-            Assert.Null(submission.StateAuxiliaryIdentifier);
+            Assert.Null(submission.CertNo);
+            Assert.Null(submission.StateAuxiliaryId);
             Assert.Null(submission.NCHSIdentifier);
 
             MessageParseException ex = Assert.Throws<MessageParseException>(() => BaseMessage.Parse(FixtureStream("fixtures/json/EmptySubmission.json")));
@@ -105,9 +105,9 @@ namespace VRDR.Tests
             Assert.Equal("nightingale", errMsg.MessageDestination);
             Assert.Equal("a9d66d2e-2480-4e8d-bab3-4e4c761da1b7", errMsg.FailedMessageId);
             Assert.Equal("2018MA000001", errMsg.NCHSIdentifier);
-            Assert.Equal((uint)1, errMsg.CertificateNumber);
+            Assert.Equal((uint)1, errMsg.CertNo);
             Assert.Equal((uint)2018, errMsg.DeathYear);
-            Assert.Equal("42", errMsg.StateAuxiliaryIdentifier);
+            Assert.Equal("42", errMsg.StateAuxiliaryId);
 
             ex = Assert.Throws<MessageParseException>(() => BaseMessage.Parse<AcknowledgementMessage>(FixtureStream("fixtures/json/DeathRecordSubmissionMessage.json")));
             Assert.Equal("The supplied message was of type VRDR.DeathRecordSubmissionMessage, expected VRDR.AcknowledgementMessage or a subclass", ex.Message);
@@ -118,15 +118,15 @@ namespace VRDR.Tests
         {
             DeathRecordSubmissionMessage submission = new DeathRecordSubmissionMessage();
             submission.DeathRecord = new DeathRecord();
-            submission.CertificateNumber = 42;
-            submission.StateAuxiliaryIdentifier = "identifier";
+            submission.CertNo = 42;
+            submission.StateAuxiliaryId = "identifier";
             Bundle submissionBundle = (Bundle)submission;
 
             DeathRecordSubmissionMessage parsed = BaseMessage.Parse<DeathRecordSubmissionMessage>(submissionBundle);
             Assert.Equal(submission.DeathRecord.ToJson(), parsed.DeathRecord.ToJson());
             Assert.Equal(submission.MessageType, parsed.MessageType);
-            Assert.Equal(submission.CertificateNumber, parsed.CertificateNumber);
-            Assert.Equal(submission.StateAuxiliaryIdentifier, parsed.StateAuxiliaryIdentifier);
+            Assert.Equal(submission.CertNo, parsed.CertNo);
+            Assert.Equal(submission.StateAuxiliaryId, parsed.StateAuxiliaryId);
             Assert.Equal(submission.NCHSIdentifier, parsed.NCHSIdentifier);
         }
 
@@ -140,8 +140,8 @@ namespace VRDR.Tests
             Assert.NotNull(submission.MessageTimestamp);
             Assert.Null(submission.MessageSource);
             Assert.NotNull(submission.MessageId);
-            Assert.Null(submission.CertificateNumber);
-            Assert.Null(submission.StateAuxiliaryIdentifier);
+            Assert.Null(submission.CertNo);
+            Assert.Null(submission.StateAuxiliaryId);
             Assert.Null(submission.NCHSIdentifier);
         }
 
@@ -152,17 +152,17 @@ namespace VRDR.Tests
             Assert.NotNull(update.DeathRecord);
             Assert.Equal("2019-02-20T16:48:06-05:00", update.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_update", update.MessageType);
-            Assert.Equal((uint)182, update.CertificateNumber);
+            Assert.Equal((uint)182, update.CertNo);
             Assert.Equal((uint)2019, update.DeathYear);
-            Assert.Equal("000000000042", update.StateAuxiliaryIdentifier);
+            Assert.Equal("000000000042", update.StateAuxiliaryId);
             Assert.Equal("2019YC000182", update.NCHSIdentifier);
 
             update = new DeathRecordUpdateMessage((DeathRecord)JSONRecords[1]); // no ids in this death record (except jurisdiction id which is required)
             Assert.NotNull(update.DeathRecord);
             Assert.Equal("2019-02-20T16:48:06-05:00", update.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_update", update.MessageType);
-            Assert.Null(update.CertificateNumber);
-            Assert.Null(update.StateAuxiliaryIdentifier);
+            Assert.Null(update.CertNo);
+            Assert.Null(update.StateAuxiliaryId);
             Assert.Null(update.NCHSIdentifier);
         }
 
@@ -173,15 +173,15 @@ namespace VRDR.Tests
             Assert.Equal("2018-02-20T16:48:06-05:00", update.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_update", update.MessageType);
             Assert.Equal("2018MA000001", update.NCHSIdentifier);
-            Assert.Equal((uint)1, update.CertificateNumber);
+            Assert.Equal((uint)1, update.CertNo);
             Assert.Equal((uint)2018, update.DeathYear);
-            Assert.Equal("42", update.StateAuxiliaryIdentifier);
+            Assert.Equal("42", update.StateAuxiliaryId);
 
             update = BaseMessage.Parse<DeathRecordUpdateMessage>(FixtureStream("fixtures/json/DeathRecordUpdateNoIdentifiers.json"));
             Assert.Equal("2018-02-20T16:48:06-05:00", update.DeathRecord.DateOfDeathPronouncement);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_update", update.MessageType);
-            Assert.Null(update.CertificateNumber);
-            Assert.Null(update.StateAuxiliaryIdentifier);
+            Assert.Null(update.CertNo);
+            Assert.Null(update.StateAuxiliaryId);
             Assert.Null(update.NCHSIdentifier);
         }
 
@@ -190,15 +190,15 @@ namespace VRDR.Tests
         {
             DeathRecordUpdateMessage update = new DeathRecordUpdateMessage();
             update.DeathRecord = new DeathRecord();
-            update.CertificateNumber = 42;
-            update.StateAuxiliaryIdentifier = "identifier";
+            update.CertNo = 42;
+            update.StateAuxiliaryId = "identifier";
             Bundle updateBundle = (Bundle)update;
 
             DeathRecordUpdateMessage parsed = BaseMessage.Parse<DeathRecordUpdateMessage>(updateBundle);
             Assert.Equal(update.DeathRecord.ToJson(), parsed.DeathRecord.ToJson());
             Assert.Equal(update.MessageType, parsed.MessageType);
-            Assert.Equal(update.CertificateNumber, parsed.CertificateNumber);
-            Assert.Equal(update.StateAuxiliaryIdentifier, parsed.StateAuxiliaryIdentifier);
+            Assert.Equal(update.CertNo, parsed.CertNo);
+            Assert.Equal(update.StateAuxiliaryId, parsed.StateAuxiliaryId);
             Assert.Equal(update.NCHSIdentifier, parsed.NCHSIdentifier);
         }
 
@@ -211,8 +211,8 @@ namespace VRDR.Tests
             Assert.Equal(submission.MessageId, ack.AckedMessageId);
             Assert.Equal(submission.MessageSource, ack.MessageDestination);
             Assert.Equal(submission.MessageDestination, ack.MessageSource);
-            Assert.Equal(submission.StateAuxiliaryIdentifier, ack.StateAuxiliaryIdentifier);
-            Assert.Equal(submission.CertificateNumber, ack.CertificateNumber);
+            Assert.Equal(submission.StateAuxiliaryId, ack.StateAuxiliaryId);
+            Assert.Equal(submission.CertNo, ack.CertNo);
             Assert.Equal(submission.NCHSIdentifier, ack.NCHSIdentifier);
 
             submission = null;
@@ -221,8 +221,8 @@ namespace VRDR.Tests
             Assert.Null(ack.AckedMessageId);
             Assert.Null(ack.MessageDestination);
             Assert.Null(ack.MessageSource);
-            Assert.Null(ack.CertificateNumber);
-            Assert.Null(ack.StateAuxiliaryIdentifier);
+            Assert.Null(ack.CertNo);
+            Assert.Null(ack.StateAuxiliaryId);
             Assert.Null(ack.NCHSIdentifier);
 
             submission = new DeathRecordSubmissionMessage();
@@ -231,8 +231,8 @@ namespace VRDR.Tests
             Assert.Equal(submission.MessageId, ack.AckedMessageId);
             Assert.Equal(submission.MessageSource, ack.MessageDestination);
             Assert.Equal(submission.MessageDestination, ack.MessageSource);
-            Assert.Equal(submission.StateAuxiliaryIdentifier, ack.StateAuxiliaryIdentifier);
-            Assert.Equal(submission.CertificateNumber, ack.CertificateNumber);
+            Assert.Equal(submission.StateAuxiliaryId, ack.StateAuxiliaryId);
+            Assert.Equal(submission.CertNo, ack.CertNo);
             Assert.Equal(submission.NCHSIdentifier, ack.NCHSIdentifier);
         }
 
@@ -244,9 +244,9 @@ namespace VRDR.Tests
             Assert.Equal("a9d66d2e-2480-4e8d-bab3-4e4c761da1b7", ack.AckedMessageId);
             Assert.Equal("nightingale", ack.MessageDestination);
             Assert.Equal("2018MA000001", ack.NCHSIdentifier);
-            Assert.Equal((uint)1, ack.CertificateNumber);
+            Assert.Equal((uint)1, ack.CertNo);
             Assert.Equal((uint)2018, ack.DeathYear);
-            Assert.Equal("42", ack.StateAuxiliaryIdentifier);
+            Assert.Equal("42", ack.StateAuxiliaryId);
         }
 
         [Fact]
@@ -257,9 +257,9 @@ namespace VRDR.Tests
             Assert.Equal("a9d66d2e-2480-4e8d-bab3-4e4c761da1b7", ack.AckedMessageId);
             Assert.Equal("nightingale", ack.MessageDestination);
             Assert.Equal("2018MA000001", ack.NCHSIdentifier);
-            Assert.Equal((uint)1, ack.CertificateNumber);
+            Assert.Equal((uint)1, ack.CertNo);
             Assert.Equal((uint)2018, ack.DeathYear);
-            Assert.Equal("42", ack.StateAuxiliaryIdentifier);
+            Assert.Equal("42", ack.StateAuxiliaryId);
         }
 
         [Fact]
@@ -271,9 +271,9 @@ namespace VRDR.Tests
             Assert.Equal("a9d66d2e-2480-4e8d-bab3-4e4c761da1b7", ack.AckedMessageId);
             Assert.Equal("nightingale", ack.MessageDestination);
             Assert.Equal("2018MA000001", ack.NCHSIdentifier);
-            Assert.Equal((uint)1, ack.CertificateNumber);
+            Assert.Equal((uint)1, ack.CertNo);
             Assert.Equal((uint)2018, ack.DeathYear);
-            Assert.Equal("42", ack.StateAuxiliaryIdentifier);
+            Assert.Equal("42", ack.StateAuxiliaryId);
         }
 
         [Fact]
@@ -289,9 +289,9 @@ namespace VRDR.Tests
 
             Assert.Equal(CauseOfDeathCodingMessage.MESSAGE_TYPE, message.MessageType);
             Assert.Equal("destination", message.MessageDestination);
-            Assert.Equal((uint)1, message.CertificateNumber);
+            Assert.Equal((uint)1, message.CertNo);
             Assert.Equal((uint)2018, message.DeathYear);
-            Assert.Equal("42", message.StateAuxiliaryIdentifier);
+            Assert.Equal("42", message.StateAuxiliaryId);
             Assert.Equal("2018MA000001", message.NCHSIdentifier);
             Assert.Equal("A04.7", message.UnderlyingCauseOfDeath);
             var recordAxisCodes = message.CauseOfDeathRecordAxis;
@@ -345,9 +345,9 @@ namespace VRDR.Tests
 
             Assert.Equal(DemographicsCodingMessage.MESSAGE_TYPE, message.MessageType);
             Assert.Equal("destination", message.MessageDestination);
-            Assert.Equal((uint)1, message.CertificateNumber);
+            Assert.Equal((uint)1, message.CertNo);
             Assert.Equal((uint)2018, message.DeathYear);
-            Assert.Equal("42", message.StateAuxiliaryIdentifier);
+            Assert.Equal("42", message.StateAuxiliaryId);
             Assert.Equal("2018MA000001", message.NCHSIdentifier);
             var ethnicity = message.Ethnicity;
             Assert.Equal(2, ethnicity.Count);
@@ -378,9 +378,9 @@ namespace VRDR.Tests
             CauseOfDeathCodingUpdateMessage message = BaseMessage.Parse<CauseOfDeathCodingUpdateMessage>(FixtureStream("fixtures/json/CauseOfDeathCodingUpdateMessage.json"));
             Assert.Equal(CauseOfDeathCodingUpdateMessage.MESSAGE_TYPE, message.MessageType);
             Assert.Equal("destination", message.MessageDestination);
-            Assert.Equal((uint)1, message.CertificateNumber);
+            Assert.Equal((uint)1, message.CertNo);
             Assert.Equal((uint)2018, message.DeathYear);
-            Assert.Equal("42", message.StateAuxiliaryIdentifier);
+            Assert.Equal("42", message.StateAuxiliaryId);
             Assert.Equal("2018MA000001", message.NCHSIdentifier);
             Assert.Equal("A04.7", message.UnderlyingCauseOfDeath);
             var recordAxisCodes = message.CauseOfDeathRecordAxis;
@@ -406,9 +406,9 @@ namespace VRDR.Tests
             DemographicsCodingUpdateMessage message = BaseMessage.Parse<DemographicsCodingUpdateMessage>(FixtureStream("fixtures/json/DemographicsCodingUpdateMessage.json"));
             Assert.Equal(DemographicsCodingUpdateMessage.MESSAGE_TYPE, message.MessageType);
             Assert.Equal("destination", message.MessageDestination);
-            Assert.Equal((uint)1, message.CertificateNumber);
+            Assert.Equal((uint)1, message.CertNo);
             Assert.Equal((uint)2018, message.DeathYear);
-            Assert.Equal("42", message.StateAuxiliaryIdentifier);
+            Assert.Equal("42", message.StateAuxiliaryId);
             Assert.Equal("2018MA000001", message.NCHSIdentifier);
             var ethnicity = message.Ethnicity;
             Assert.Equal(2, ethnicity.Count);
@@ -433,13 +433,13 @@ namespace VRDR.Tests
             Assert.Equal(CauseOfDeathCodingMessage.MESSAGE_TYPE, message.MessageType);
             Assert.Equal("destination", message.MessageDestination);
 
-            Assert.Null(message.CertificateNumber);
-            message.CertificateNumber = 10;
-            Assert.Equal((uint)10, message.CertificateNumber);
+            Assert.Null(message.CertNo);
+            message.CertNo = 10;
+            Assert.Equal((uint)10, message.CertNo);
 
-            Assert.Null(message.StateAuxiliaryIdentifier);
-            message.StateAuxiliaryIdentifier = "id101010";
-            Assert.Equal("id101010", message.StateAuxiliaryIdentifier);
+            Assert.Null(message.StateAuxiliaryId);
+            message.StateAuxiliaryId = "id101010";
+            Assert.Equal("id101010", message.StateAuxiliaryId);
 
             Assert.Null(message.DeathYear);
             message.DeathYear = 2019;
@@ -501,9 +501,9 @@ namespace VRDR.Tests
             message.OtherSpecifiedPlace = "Unique Location";
             Assert.Equal("Unique Location", message.OtherSpecifiedPlace);
 
-            Assert.Null(message.DeathJurisdictionID);
-            message.DeathJurisdictionID = "NH";
-            Assert.Equal("NH", message.DeathJurisdictionID);
+            Assert.Null(message.JurisdictionId);
+            message.JurisdictionId = "NH";
+            Assert.Equal("NH", message.JurisdictionId);
             Assert.Equal("2019NH000010", message.NCHSIdentifier);
 
             Assert.Null(message.IntentionalReject);
@@ -570,13 +570,13 @@ namespace VRDR.Tests
             Assert.Equal(DemographicsCodingMessage.MESSAGE_TYPE, message.MessageType);
             Assert.Equal("destination", message.MessageDestination);
 
-            Assert.Null(message.CertificateNumber);
-            message.CertificateNumber = 10;
-            Assert.Equal((uint)10, message.CertificateNumber);
+            Assert.Null(message.CertNo);
+            message.CertNo = 10;
+            Assert.Equal((uint)10, message.CertNo);
 
-            Assert.Null(message.StateAuxiliaryIdentifier);
-            message.StateAuxiliaryIdentifier = "id101010";
-            Assert.Equal("id101010", message.StateAuxiliaryIdentifier);
+            Assert.Null(message.StateAuxiliaryId);
+            message.StateAuxiliaryId = "id101010";
+            Assert.Equal("id101010", message.StateAuxiliaryId);
 
             Assert.Null(message.DeathYear);
             message.DeathYear = 2019;
@@ -626,9 +626,9 @@ namespace VRDR.Tests
             message.ACMESystemRejectCodes = CodingResponseMessage.ACMESystemRejectEnum.ACMEReject;
             Assert.Equal(CodingResponseMessage.ACMESystemRejectEnum.ACMEReject, message.ACMESystemRejectCodes);
 
-            Assert.Null(message.DeathJurisdictionID);
-            message.DeathJurisdictionID = "NH";
-            Assert.Equal("NH", message.DeathJurisdictionID);
+            Assert.Null(message.JurisdictionId);
+            message.JurisdictionId = "NH";
+            Assert.Equal("NH", message.JurisdictionId);
             Assert.Equal("2019NH000010", message.NCHSIdentifier);
 
             Assert.Null(message.IntentionalReject);
@@ -693,21 +693,21 @@ namespace VRDR.Tests
             Assert.Equal(CauseOfDeathCodingUpdateMessage.MESSAGE_TYPE, message.MessageType);
             Assert.Equal("destination", message.MessageDestination);
 
-            Assert.Null(message.CertificateNumber);
-            message.CertificateNumber = 10;
-            Assert.Equal((uint)10, message.CertificateNumber);
+            Assert.Null(message.CertNo);
+            message.CertNo = 10;
+            Assert.Equal((uint)10, message.CertNo);
 
-            Assert.Null(message.StateAuxiliaryIdentifier);
-            message.StateAuxiliaryIdentifier = "id101010";
-            Assert.Equal("id101010", message.StateAuxiliaryIdentifier);
+            Assert.Null(message.StateAuxiliaryId);
+            message.StateAuxiliaryId = "id101010";
+            Assert.Equal("id101010", message.StateAuxiliaryId);
 
             Assert.Null(message.DeathYear);
             message.DeathYear = 2019;
             Assert.Equal((uint)2019, message.DeathYear);
 
-            Assert.Null(message.DeathJurisdictionID);
-            message.DeathJurisdictionID = "NH";
-            Assert.Equal("NH", message.DeathJurisdictionID);
+            Assert.Null(message.JurisdictionId);
+            message.JurisdictionId = "NH";
+            Assert.Equal("NH", message.JurisdictionId);
             Assert.Equal("2019NH000010", message.NCHSIdentifier);
 
             Assert.Null(message.UnderlyingCauseOfDeath);
@@ -756,21 +756,21 @@ namespace VRDR.Tests
             Assert.Equal(DemographicsCodingUpdateMessage.MESSAGE_TYPE, message.MessageType);
             Assert.Equal("destination", message.MessageDestination);
 
-            Assert.Null(message.CertificateNumber);
-            message.CertificateNumber = 10;
-            Assert.Equal((uint)10, message.CertificateNumber);
+            Assert.Null(message.CertNo);
+            message.CertNo = 10;
+            Assert.Equal((uint)10, message.CertNo);
 
-            Assert.Null(message.StateAuxiliaryIdentifier);
-            message.StateAuxiliaryIdentifier = "id101010";
-            Assert.Equal("id101010", message.StateAuxiliaryIdentifier);
+            Assert.Null(message.StateAuxiliaryId);
+            message.StateAuxiliaryId = "id101010";
+            Assert.Equal("id101010", message.StateAuxiliaryId);
 
             Assert.Null(message.DeathYear);
             message.DeathYear = 2019;
             Assert.Equal((uint)2019, message.DeathYear);
 
-            Assert.Null(message.DeathJurisdictionID);
-            message.DeathJurisdictionID = "NH";
-            Assert.Equal("NH", message.DeathJurisdictionID);
+            Assert.Null(message.JurisdictionId);
+            message.JurisdictionId = "NH";
+            Assert.Equal("NH", message.JurisdictionId);
             Assert.Equal("2019NH000010", message.NCHSIdentifier);
 
             Assert.Empty(message.Ethnicity);
@@ -819,12 +819,12 @@ namespace VRDR.Tests
         {
             DeathRecordVoidMessage message = new DeathRecordVoidMessage();
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_void", message.MessageType);
-            Assert.Null(message.CertificateNumber);
-            message.CertificateNumber = 11;
-            Assert.Equal((uint)11, message.CertificateNumber);
-            Assert.Null(message.StateAuxiliaryIdentifier);
-            message.StateAuxiliaryIdentifier = "bar";
-            Assert.Equal("bar", message.StateAuxiliaryIdentifier);
+            Assert.Null(message.CertNo);
+            message.CertNo = 11;
+            Assert.Equal((uint)11, message.CertNo);
+            Assert.Null(message.StateAuxiliaryId);
+            message.StateAuxiliaryId = "bar";
+            Assert.Equal("bar", message.StateAuxiliaryId);
             Assert.Null(message.NCHSIdentifier);
             Assert.Null(message.BlockCount);
             message.BlockCount = 100;
@@ -840,8 +840,8 @@ namespace VRDR.Tests
             Assert.Equal(voidMessage.MessageId, ack.AckedMessageId);
             Assert.Equal(voidMessage.MessageSource, ack.MessageDestination);
             Assert.Equal(voidMessage.MessageDestination, ack.MessageSource);
-            Assert.Equal(voidMessage.StateAuxiliaryIdentifier, ack.StateAuxiliaryIdentifier);
-            Assert.Equal(voidMessage.CertificateNumber, ack.CertificateNumber);
+            Assert.Equal(voidMessage.StateAuxiliaryId, ack.StateAuxiliaryId);
+            Assert.Equal(voidMessage.CertNo, ack.CertNo);
             Assert.Equal(voidMessage.NCHSIdentifier, ack.NCHSIdentifier);
             Assert.Equal(voidMessage.BlockCount, ack.BlockCount);
 
@@ -851,8 +851,8 @@ namespace VRDR.Tests
             Assert.Null(ack.AckedMessageId);
             Assert.Null(ack.MessageDestination);
             Assert.Null(ack.MessageSource);
-            Assert.Null(ack.CertificateNumber);
-            Assert.Null(ack.StateAuxiliaryIdentifier);
+            Assert.Null(ack.CertNo);
+            Assert.Null(ack.StateAuxiliaryId);
             Assert.Null(ack.NCHSIdentifier);
             Assert.Null(ack.BlockCount);
 
@@ -862,8 +862,8 @@ namespace VRDR.Tests
             Assert.Equal(voidMessage.MessageId, ack.AckedMessageId);
             Assert.Equal(voidMessage.MessageSource, ack.MessageDestination);
             Assert.Equal(voidMessage.MessageDestination, ack.MessageSource);
-            Assert.Equal(voidMessage.StateAuxiliaryIdentifier, ack.StateAuxiliaryIdentifier);
-            Assert.Equal(voidMessage.CertificateNumber, ack.CertificateNumber);
+            Assert.Equal(voidMessage.StateAuxiliaryId, ack.StateAuxiliaryId);
+            Assert.Equal(voidMessage.CertNo, ack.CertNo);
             Assert.Equal(voidMessage.NCHSIdentifier, ack.NCHSIdentifier);
             Assert.Equal(voidMessage.BlockCount, ack.BlockCount);
         }
@@ -873,17 +873,17 @@ namespace VRDR.Tests
         {
             DeathRecordVoidMessage message = BaseMessage.Parse<DeathRecordVoidMessage>(FixtureStream("fixtures/json/DeathRecordVoidMessage.json"));
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_void", message.MessageType);
-            Assert.Equal((uint)1, message.CertificateNumber);
+            Assert.Equal((uint)1, message.CertNo);
             Assert.Equal((uint)10, message.BlockCount);
-            Assert.Equal("42", message.StateAuxiliaryIdentifier);
+            Assert.Equal("42", message.StateAuxiliaryId);
             Assert.Equal("2018MA000001", message.NCHSIdentifier);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", message.MessageDestination);
             Assert.Equal("nightingale", message.MessageSource);
 
             message = BaseMessage.Parse<DeathRecordVoidMessage>(FixtureStream("fixtures/json/DeathRecordVoidMessageNoIdentifiers.json"));
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_void", message.MessageType);
-            Assert.Null(message.CertificateNumber);
-            Assert.Null(message.StateAuxiliaryIdentifier);
+            Assert.Null(message.CertNo);
+            Assert.Null(message.StateAuxiliaryId);
             Assert.Null(message.NCHSIdentifier);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", message.MessageDestination);
             Assert.Equal("nightingale", message.MessageSource);
@@ -894,16 +894,16 @@ namespace VRDR.Tests
         {
             DeathRecordVoidMessage message = new DeathRecordVoidMessage((DeathRecord)XMLRecords[0]);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_void", message.MessageType);
-            Assert.Equal((uint)182, message.CertificateNumber);
-            Assert.Equal("000000000042", message.StateAuxiliaryIdentifier);
+            Assert.Equal((uint)182, message.CertNo);
+            Assert.Equal("000000000042", message.StateAuxiliaryId);
             Assert.Equal("2019YC000182", message.NCHSIdentifier);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", message.MessageDestination);
             Assert.Null(message.MessageSource);
 
             message = new DeathRecordVoidMessage(null);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission_void", message.MessageType);
-            Assert.Null(message.CertificateNumber);
-            Assert.Null(message.StateAuxiliaryIdentifier);
+            Assert.Null(message.CertNo);
+            Assert.Null(message.StateAuxiliaryId);
             Assert.Null(message.NCHSIdentifier);
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", message.MessageDestination);
             Assert.Null(message.MessageSource);
@@ -935,9 +935,9 @@ namespace VRDR.Tests
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", responseMsg.MessageDestination);
             Assert.Equal("nightingale", responseMsg.MessageSource);
             Assert.Equal("761dca08-259b-4dcd-aeb7-bb3c73fa30f2", responseMsg.FailedMessageId);
-            Assert.Null(responseMsg.CertificateNumber);
+            Assert.Null(responseMsg.CertNo);
             Assert.Null(responseMsg.NCHSIdentifier);
-            Assert.Null(responseMsg.StateAuxiliaryIdentifier);
+            Assert.Null(responseMsg.StateAuxiliaryId);
 
             ex = Assert.Throws<MessageParseException>(() => BaseMessage.Parse(FixtureStream("fixtures/json/MissingMessageType.json")));
             Assert.Equal("Message type was missing from MessageHeader", ex.Message);
@@ -945,9 +945,9 @@ namespace VRDR.Tests
             Assert.Equal("http://nchs.cdc.gov/vrdr_submission", responseMsg.MessageDestination);
             Assert.Equal("nightingale", responseMsg.MessageSource);
             Assert.Equal("761dca08-259b-4dcd-aeb7-bb3c73fa30f2", responseMsg.FailedMessageId);
-            Assert.Null(responseMsg.CertificateNumber);
+            Assert.Null(responseMsg.CertNo);
             Assert.Null(responseMsg.NCHSIdentifier);
-            Assert.Null(responseMsg.StateAuxiliaryIdentifier);
+            Assert.Null(responseMsg.StateAuxiliaryId);
 
             ex = Assert.Throws<MessageParseException>(() => BaseMessage.Parse(FixtureStream("fixtures/json/EmptyMessage.json")));
             Assert.Equal("Failed to find a Bundle Entry containing a Resource of type MessageHeader", ex.Message);
@@ -955,9 +955,9 @@ namespace VRDR.Tests
             Assert.Null(responseMsg.MessageDestination);
             Assert.Null(responseMsg.MessageSource);
             Assert.Null(responseMsg.FailedMessageId);
-            Assert.Null(responseMsg.CertificateNumber);
+            Assert.Null(responseMsg.CertNo);
             Assert.Null(responseMsg.NCHSIdentifier);
-            Assert.Null(responseMsg.StateAuxiliaryIdentifier);
+            Assert.Null(responseMsg.StateAuxiliaryId);
 
             ex = Assert.Throws<MessageParseException>(() => BaseMessage.Parse(FixtureStream("fixtures/json/Empty.json")));
             Assert.Equal("The FHIR Bundle must be of type message, not null", ex.Message);
@@ -965,9 +965,9 @@ namespace VRDR.Tests
             Assert.Null(responseMsg.MessageDestination);
             Assert.Null(responseMsg.MessageSource);
             Assert.Null(responseMsg.FailedMessageId);
-            Assert.Null(responseMsg.CertificateNumber);
+            Assert.Null(responseMsg.CertNo);
             Assert.Null(responseMsg.NCHSIdentifier);
-            Assert.Null(responseMsg.StateAuxiliaryIdentifier);
+            Assert.Null(responseMsg.StateAuxiliaryId);
         }
 
         [Fact]
@@ -978,8 +978,8 @@ namespace VRDR.Tests
             Assert.Equal("http://nchs.cdc.gov/vrdr_extraction_error", err.MessageType);
             Assert.Equal(submission.MessageId, err.FailedMessageId);
             Assert.Equal(submission.MessageSource, err.MessageDestination);
-            Assert.Equal(submission.StateAuxiliaryIdentifier, err.StateAuxiliaryIdentifier);
-            Assert.Equal(submission.CertificateNumber, err.CertificateNumber);
+            Assert.Equal(submission.StateAuxiliaryId, err.StateAuxiliaryId);
+            Assert.Equal(submission.CertNo, err.CertNo);
             Assert.Equal(submission.NCHSIdentifier, err.NCHSIdentifier);
             Assert.Empty(err.Issues);
             var issues = new List<Issue>();
@@ -1002,8 +1002,8 @@ namespace VRDR.Tests
             Assert.Equal("http://nchs.cdc.gov/vrdr_extraction_error", err.MessageType);
             Assert.Null(err.FailedMessageId);
             Assert.Null(err.MessageDestination);
-            Assert.Null(err.CertificateNumber);
-            Assert.Null(err.StateAuxiliaryIdentifier);
+            Assert.Null(err.CertNo);
+            Assert.Null(err.StateAuxiliaryId);
             Assert.Null(err.NCHSIdentifier);
             Assert.Empty(err.Issues);
         }
@@ -1013,8 +1013,8 @@ namespace VRDR.Tests
         {
             ExtractionErrorMessage err = BaseMessage.Parse<ExtractionErrorMessage>(FixtureStream("fixtures/json/ExtractionErrorMessage.json"));
             Assert.Equal("http://nchs.cdc.gov/vrdr_extraction_error", err.MessageType);
-            Assert.Equal((uint)1, err.CertificateNumber);
-            Assert.Equal("42", err.StateAuxiliaryIdentifier);
+            Assert.Equal((uint)1, err.CertNo);
+            Assert.Equal("42", err.StateAuxiliaryId);
             Assert.Equal("2018MA000001", err.NCHSIdentifier);
             var issues = err.Issues;
             Assert.Equal(2, (int)issues.Count);